+++ examples/cpp/auth/helper.h
@@ -1,0 +1,26 @@
+/*
+ *
+ * Copyright 2024 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#ifndef GRPC_EXAMPLES_CPP_AUTH_HELPER_H_
+#define GRPC_EXAMPLES_CPP_AUTH_HELPER_H_
+
+#include <string>
+
+std::string LoadStringFromFile(std::string path);
+
+#endif  // GRPC_EXAMPLES_CPP_AUTH_HELPER_H_
